Establish dependency between generated sources and their tasks
If we set the source directory to a Property<File> instead of a blank File,
then (if the Property depends on a task) Gradle can infer the task which needs
to be ran to generate the sources.

This is sufficient to establish a dependency between e.g. the Java plugin's
sourcesJar task and our preprocessJava task.

Still keeping all the explicitly declared dependencies though cause some of
those are for ForgeGradle tasks and I highly doubt ForgeGradle respects the
dependencies on the source sets.