Combine Java and Kotlin preprocess tasks into a single task
Cause both can be remapped at the same time and doing it in one go allows us to
save on overhead from parsing the classpath twice.

This also removes the requirement for the Java task to wait on compileKotlin to
get the Kotlin classes to link against because the preprocess task now has
access to both at the same time. Note that while the requirement is dropped, the
task dependency is still present and will be removed in the next commit.