Remove task dependency of preprocess tasks on compile tasks
Allowing them to run in parallel, potentially halving overall build time in the
best case.
This dependency used to be required so the Java preprocess task had access to
the compiled Kotlin class file, but the parent commit merges both preprocess
task into one, so remap can simply compile everything it needs by itself (it
would have had to compile everything at least once anyway, just split over two
tasks).

We use the classpath of the compileKotlin task when available because the
compileJava classpath contains an implicit dependency on compileKotlin (which is
exactly what we want to avoid). The two classpaths should otherwise be identical
with any normal project.